fixed arch problems in transaction
